# Lighter construction of finite field elements from lists