@remyoudompheng 2023-03-27
Elliptic curve isogenies provide examples for a few scenarios:
* polynomial with many zeros, most coefficients belong to the prime subfield
* polynomial with many zeros
* polynomial with few zeros

Benchmark before patch:
```
sage: p = next_prime(2**128)
sage: K = GF(p**2, 'a')
sage: E = EllipticCurve(K, [0, 1])
sage: pol = E.torsion_polynomial(97).monic()
sage: %timeit E.isogeny(pol, check=False)
1.3 s ± 4.56 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

sage: E = EllipticCurve(K, [0, K.gen()])
sage: pol = E.torsion_polynomial(97).monic()
sage: %timeit E.isogeny(pol, check=False)
2.38 s ± 6 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

sage: E = EllipticCurve(K, [K.gen(), K.gen()])
sage: pol = E.torsion_polynomial(97).monic()
sage: %timeit E.isogeny(pol, check=False)
4.78 s ± 8.36 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

After patch:
```
sage: p = next_prime(2**128)
sage: K = GF(p**2, 'a')
sage: E = EllipticCurve(K, [0, 1])
sage: pol = E.torsion_polynomial(97).monic()
sage: %timeit E.isogeny(pol, check=False)
602 ms ± 1.58 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

sage: E = EllipticCurve(K, [0, K.gen()])
sage: pol = E.torsion_polynomial(97).monic()
sage: %timeit E.isogeny(pol, check=False)
779 ms ± 4.41 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

sage: E = EllipticCurve(K, [K.gen(), K.gen()])
sage: pol = E.torsion_polynomial(97).monic()
sage: %timeit E.isogeny(pol, check=False)
1.48 s ± 13.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```